test_task/requirements.txt
pytest==7.4.3
allure-pytest==2.13.2

selenium~=4.14.0